+++ app/controllers/posts_controller.rb
@@ -6,31 +6,39 @@ class PostsController < ApplicationController
     def create
         # Creates an instance of post
         @post = Post.new(post_params)
-        # Save it in DB
-        if @post.save
-            render json: {
-                message: 'Post created successfully',
-                post: {
-                    id: @post.id,
-                    title: @post.title,
-                    url: @post.url,
-                    body: @post.body,
-                    creator_id: @post.user.id,
-                    community_id: @post.community.id,
-                    created_at: @post.created_at,
-                    updated_at: @post.updated_at
-                }
-            }, status: :created
-        else
-            render json: {
-                errors: @post.errors.full_messages
-            }, status: :unprocessable_entity
+
+        if @current_user != nil
+          @post.user_id = @current_user.id
+        elsif api_key.nil?
+          render :json => { "status" => "401", "error" => "No Api key provided." }, status: :unauthorized and return
         end
+        # Save it in DB
+          if @post.save
+              render json: {
+                  message: 'Post created successfully',
+                  post: {
+                      id: @post.id,
+                      title: @post.title,
+                      url: @post.url,
+                      body: @post.body,
+                      creator_id: @post.user_id,
+                      community_id: @post.community.id,
+                      created_at: @post.created_at,
+                      updated_at: @post.updated_at
+                  }
+              }, status: :created
+          else
+              render json: {
+                  errors: @post.errors.full_messages
+              }, status: :unprocessable_entity
+              end
+
     end
 
     # GET /posts or /posts.json
     def index
         @posts = Post.all
+        authenticate_user2
 
         # Buscamos los posts por title
         if params[:search].present?
@@ -40,30 +48,20 @@ class PostsController < ApplicationController
 
         # Filtramos los posts
         if params[:filter].present?
-            if params[:user_id].present?
-                user_id = params[:user_id]
                 subs = params[:filter]
                 case subs
                     when 'subscribed'
                         @posts = Post.joins(:community, community: :subscriptions)
-                                     .where(subscriptions: { user_id: user_id })
+                                     .where(subscriptions: { user_id: @current_user.id})
                                      .order('posts.created_at DESC')
                     when 'created'
-                        @posts = Post.where( user_id: user_id )
+                        @posts = Post.where( user_id: @current_user )
                                      .order('posts.created_at DESC')
                     when 'saved'
                         @posts = Post.joins(:saved_posts)
-                                     .where(saved_posts: { user_id: user_id })
+                                     .where(saved_posts: { user_id: @current_user.id })
                                      .order('posts.created_at DESC')
                 end
-
-            else
-                #si no hay user id no podemos hacer ninguno de los filtros
-                render json: {
-                    errors: "Mismatch user_id"
-                }, status: :unprocessable_entity
-                return
-            end
         end
 
         # Ordenamos los posts
@@ -99,14 +97,22 @@ class PostsController < ApplicationController
             url: post.url,
             body: post.body,
             user_id: post.user_id,
+            user_name: User.find(post.user_id).username ,
+            user_avatar: User.find(post.user_id).user_avatar.attached? ? url_for( User.find(post.user_id).user_avatar) : nil,
+            user_saved: SavedPost.find_by(post_id: post.id, user_id: @current_user.id).present? ? true : false,
             community_id: post.community_id,
+            community_name: Community.find(post.community_id).name,
+            community_avatar: Community.find(post.community_id).community_avatar.attached? ? url_for( Community.find(post.community_id).community_avatar) : nil,
             created_at: post.created_at,
             updated_at: post.updated_at,
             num_comments: post.comments.count,
             likes: {
                 positive: post.post_likes.where(positive: true).count || 0,
-                negative: post.post_likes.where(positive: false).count || 0
+                negative: post.post_likes.where(positive: false).count || 0,
+                user_like: @current_user ? post.post_likes.where(user: @current_user, positive: true).exists? : nil
             }
+
+
         }
         end
 
@@ -158,11 +164,14 @@ class PostsController < ApplicationController
     # DELETE /posts/1
     def destroy
       @post = Post.find(params[:id])
-
-      if @post.destroy
-        render json: { message: 'Post deleted successfully' }, status: :ok
+      if @post.user_id == @current_user.id
+        if @post.destroy
+          render json: { message: 'Post deleted successfully' }, status: :ok
+        else
+          render json: { errors: @post.errors.full_messages }, status: :unprocessable_entity
+        end
       else
-        render json: { errors: @post.errors.full_messages }, status: :unprocessable_entity
+        render json: { message: 'This user is not the creator' }, status: :unprocessable_entity
       end
     end
 
@@ -170,8 +179,8 @@ class PostsController < ApplicationController
     # PUT /posts/1
     def update
         @post = Post.find(params[:id])
-
-        if @post.update(post_params)
+        if @post.user_id == @current_user.id
+          if @post.update(post_params)
             render json: {
                 message: 'Post updated successfully',
                 post: {
@@ -190,20 +199,22 @@ class PostsController < ApplicationController
                     }
                 }
             }, status: :ok
+          else
+              render json: {
+              errors: @post.errors.full_messages
+              }, status: :unprocessable_entity
+          end
         else
-            render json: {
-            errors: @post.errors.full_messages
-            }, status: :unprocessable_entity
+          render json: { message: 'This user is not the creator' }, status: :unprocessable_entity
         end
     end
-
 
 
 
     private
         # Only allow a list of trusted parameters through.
         def post_params
-            params.require(:post).permit(:title, :url, :body, :user_id, :community_id)
+            params.require(:post).permit(:title, :url, :body, :community_id)
         end
 
 
@@ -225,4 +236,11 @@ class PostsController < ApplicationController
         def token_valid?(token)
             !token.nil? && !token.empty? && User.exists?(api_key: token)
         end
+        def authenticate_user2
+          token = extract_token_from_request
+          if token_valid?(token)
+            @current_user = User.find_by(api_key: token)
+          end
+        end
+
 end
